tests run now
