--- ui/test-shim.js
@@ -1,1 +1,0 @@
-require('mutationobserver-shim');
